\name{FM.data}
\alias{FM.data}
\title{Finds all intervals compatible under frequent monitoring assumption in a synthetic dataset}
\usage{
  FM.data(simDataList, u)
}
\arguments{
  \item{simDataList}{A list of synthetic observed datasets,
  returned by \code{\link{makedata.simple}}}

  \item{u}{The index of the desired entry of simDataList}
}
\value{
  A list containing information about each type of FM event
}
\description{
  \code{FM.data} takes a simulated dataset in the format
  generated by \code{\link{makedata.simple}} and finds the
  subset of observation intervals where at most one event
  occurred. This is necessary for computation of the log
  likelihood in \code{\link{logFM}}
}

