\name{MSTEP}
\alias{MSTEP}
\title{Execute a Newton-Raphson step in M-step of the EM algorithm}
\usage{
  MSTEP(matrix, betaVec, num.patients, patients.design)
}
\arguments{
  \item{matrix}{A matrix in the format returned by
  \code{\link{ESTEP}} or \code{\link{ESTEP.slow}}}

  \item{betaVec}{A vector of regression coefficients}

  \item{num.patients}{An integer, the number of unique
  patients}

  \item{patients.design}{A design matrix in the format
  generated by \code{\link{PatientDesignExample}}}
}
\value{
  An updated coefficient vector after one Newton-Raphson
  step
}
\description{
  \code{MSTEP} executes one iteration of a Newton-Raphson
  algorithm as part of the maximization (M-step) of the EM
  algorithm. Given the matrix of expected sufficient
  statistics returned by \code{\link{ESTEP}}, this function
  uses closed form gradient and hessian expressions to
  efficiently optimize the current settings of the
  coefficients beta. This is called up to 10 times per
  M-step within \code{\link{EM.run}}
}

